fix meteor hanging during build
